+++ app/graphql/queries/stack/get_stack.rb
@@ -2,10 +2,10 @@ module Queries
   module Stack
     class GetStack < Queries::BaseQuery
       argument :id, ID, required: true
-      type Types::StackType, null: true
+      type Types::Stack::StackType, null: true
 
       def resolve(id:)
-        stack_mock = {
+        {
           id: 1,
           name: 'React'
         }
