+++ app/graphql/types/company/company_type.rb
@@ -1,6 +1,6 @@
 module Types
-  module User
-    class UserType < Types::BaseObject
+  module Company
+    class CompanyType < Types::BaseObject
       field :id, ID, null: false
       field :name, String, null: true
       field :image_path, String, null: true
